# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Property Sale Screening API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/companies": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Get Companies",
        "description": "\u4f01\u696d\u30e9\u30f3\u30ad\u30f3\u30b0\u4e00\u89a7 / \u691c\u7d22\u3092\u53d6\u5f97\u3059\u308b\uff08is_current=True \u306e\u30b9\u30af\u30ea\u30fc\u30cb\u30f3\u30b0\u7d50\u679c\u304b\u3089\uff09.\n\n- q / company_name / securities_code \u304c\u6307\u5b9a\u3055\u308c\u305f\u5834\u5408\u306f\u691c\u7d22\u30b5\u30fc\u30d3\u30b9\u306b\u59d4\u8b72\u3059\u308b\u3002\n- \u3044\u305a\u308c\u3082\u7121\u3044\u5834\u5408\u306f\u5f93\u6765\u306e\u30e9\u30f3\u30ad\u30f3\u30b0\u4e00\u89a7\u3092\u8fd4\u3059\u3002",
        "operationId": "get_companies_api_v1_companies_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\u81ea\u7136\u8a00\u8a9e\u691c\u7d22\u30af\u30a8\u30ea\uff08\u6700\u5927200\u5b57\uff09\u3002LLM \u3067\u6761\u4ef6\u62bd\u51fa",
              "title": "Q"
            },
            "description": "\u81ea\u7136\u8a00\u8a9e\u691c\u7d22\u30af\u30a8\u30ea\uff08\u6700\u5927200\u5b57\uff09\u3002LLM \u3067\u6761\u4ef6\u62bd\u51fa"
          },
          {
            "name": "company_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\u4f01\u696d\u540d\uff08\u5b8c\u5168\u4e00\u81f4\uff09",
              "title": "Company Name"
            },
            "description": "\u4f01\u696d\u540d\uff08\u5b8c\u5168\u4e00\u81f4\uff09"
          },
          {
            "name": "securities_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\u8a3c\u5238\u30b3\u30fc\u30c9\uff08\u5b8c\u5168\u4e00\u81f4\uff09",
              "title": "Securities Code"
            },
            "description": "\u8a3c\u5238\u30b3\u30fc\u30c9\uff08\u5b8c\u5168\u4e00\u81f4\uff09"
          },
          {
            "name": "industry",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\u696d\u7a2e\u30d5\u30a3\u30eb\u30bf",
              "title": "Industry"
            },
            "description": "\u696d\u7a2e\u30d5\u30a3\u30eb\u30bf"
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\u6700\u4f4e\u30b9\u30b3\u30a2",
              "title": "Min Score"
            },
            "description": "\u6700\u4f4e\u30b9\u30b3\u30a2"
          },
          {
            "name": "max_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\u6700\u9ad8\u30b9\u30b3\u30a2",
              "title": "Max Score"
            },
            "description": "\u6700\u9ad8\u30b9\u30b3\u30a2"
          },
          {
            "name": "has_event",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\u30a4\u30d9\u30f3\u30c8\u30b7\u30b0\u30ca\u30eb\u306e\u6709\u7121",
              "title": "Has Event"
            },
            "description": "\u30a4\u30d9\u30f3\u30c8\u30b7\u30b0\u30ca\u30eb\u306e\u6709\u7121"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "\u30bd\u30fc\u30c8\u30ad\u30fc",
              "default": "total_score",
              "title": "Sort By"
            },
            "description": "\u30bd\u30fc\u30c8\u30ad\u30fc"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "\u30da\u30fc\u30b8\u756a\u53f7\uff081\u59cb\u307e\u308a\uff09",
              "default": 1,
              "title": "Page"
            },
            "description": "\u30da\u30fc\u30b8\u756a\u53f7\uff081\u59cb\u307e\u308a\uff09"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "1\u30da\u30fc\u30b8\u3042\u305f\u308a\u306e\u4ef6\u6570",
              "default": 20,
              "title": "Page Size"
            },
            "description": "1\u30da\u30fc\u30b8\u3042\u305f\u308a\u306e\u4ef6\u6570"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanySearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{company_id}": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Get Company Detail",
        "description": "\u4f01\u696d\u8a73\u7d30\u3092\u53d6\u5f97\u3059\u308b.",
        "operationId": "get_company_detail_api_v1_companies__company_id__get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDetailSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{company_id}/report": {
      "post": {
        "tags": [
          "companies"
        ],
        "summary": "Generate Report",
        "description": "PDF \u30ec\u30dd\u30fc\u30c8\u3092\u975e\u540c\u671f\u751f\u6210\u3059\u308b\uff08202 Accepted\uff09.",
        "operationId": "generate_report_api_v1_companies__company_id__report_post",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportGenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CompanyDetailSchema": {
        "properties": {
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id",
            "description": "\u4f01\u696dID"
          },
          "securities_code": {
            "type": "string",
            "maxLength": 10,
            "minLength": 1,
            "title": "Securities Code",
            "description": "\u8a3c\u5238\u30b3\u30fc\u30c9"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "\u4f01\u696d\u540d"
          },
          "industry": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Industry",
            "description": "\u696d\u7a2e"
          },
          "market_cap": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap",
            "description": "\u6642\u4fa1\u7dcf\u984d\uff08\u5104\u5186\uff09"
          },
          "scoring": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScoreBreakdownDetailSchema"
              },
              {
                "type": "null"
              }
            ],
            "description": "\u6700\u65b0\u30b9\u30b3\u30a2\u5185\u8a33\u3002\u30b9\u30b3\u30a2\u672a\u5b9f\u884c\u306e\u5834\u5408\u306f None"
          },
          "financial": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinancialDataDetailSchema"
              },
              {
                "type": "null"
              }
            ],
            "description": "\u6700\u65b0\u8ca1\u52d9\u30c7\u30fc\u30bf\u3002\u30c7\u30fc\u30bf\u306a\u3057\u306e\u5834\u5408\u306f None"
          },
          "signals_support": {
            "items": {
              "$ref": "#/components/schemas/QualitativeSignalDetailSchema"
            },
            "type": "array",
            "title": "Signals Support",
            "description": "\u652f\u6301\u30b7\u30b0\u30ca\u30eb\u4e00\u89a7\uff08stance=support\uff09"
          },
          "signals_counter": {
            "items": {
              "$ref": "#/components/schemas/QualitativeSignalDetailSchema"
            },
            "type": "array",
            "title": "Signals Counter",
            "description": "\u53cd\u5bfe\u30b7\u30b0\u30ca\u30eb\u4e00\u89a7\uff08stance=counter\uff09"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "securities_code",
          "name",
          "industry"
        ],
        "title": "CompanyDetailSchema",
        "description": "\u4f01\u696d\u8a73\u7d30\u30ec\u30b9\u30dd\u30f3\u30b9\uff08P002 \u4f01\u696d\u8a73\u7d30\u753b\u9762\u7528\uff09."
      },
      "CompanyRankingItemSchema": {
        "properties": {
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id",
            "description": "\u4f01\u696dID"
          },
          "securities_code": {
            "type": "string",
            "maxLength": 10,
            "minLength": 1,
            "title": "Securities Code",
            "description": "\u8a3c\u5238\u30b3\u30fc\u30c9"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "\u4f01\u696d\u540d"
          },
          "industry": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Industry",
            "description": "\u696d\u7a2e"
          },
          "market_cap": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap",
            "description": "\u6642\u4fa1\u7dcf\u984d\uff08\u5104\u5186\uff09"
          },
          "total_score": {
            "type": "number",
            "minimum": 0.0,
            "title": "Total Score",
            "description": "\u7d71\u5408\u7dcf\u5408\u30b9\u30b3\u30a2"
          },
          "structure_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Structure Score",
            "description": "\u69cb\u9020\u30b9\u30b3\u30a2\uff080\u2013100\uff09"
          },
          "event_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Event Score",
            "description": "\u30a4\u30d9\u30f3\u30c8\u30b9\u30b3\u30a2\uff080\u2013100\uff09"
          },
          "event_boost": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Boost",
            "description": "\u30a4\u30d9\u30f3\u30c8\u30d6\u30fc\u30b9\u30c8\u4fc2\u6570"
          },
          "confidence": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ConfidenceLevelEnum"
              }
            ],
            "description": "\u4fe1\u983c\u5ea6\u30ec\u30d9\u30eb"
          },
          "unrealized_gain": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unrealized Gain",
            "description": "\u542b\u307f\u76ca\uff08\u5104\u5186\uff09"
          },
          "has_event": {
            "type": "boolean",
            "title": "Has Event",
            "description": "\u30a4\u30d9\u30f3\u30c8\u30b7\u30b0\u30ca\u30eb\u304c1\u4ef6\u4ee5\u4e0a\u3042\u308b\u304b"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "securities_code",
          "name",
          "industry",
          "total_score",
          "structure_score",
          "event_score",
          "confidence",
          "has_event"
        ],
        "title": "CompanyRankingItemSchema",
        "description": "\u4f01\u696d\u30e9\u30f3\u30ad\u30f3\u30b0\u4e00\u89a7\u306e1\u30a8\u30f3\u30c8\u30ea."
      },
      "CompanySearchResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CompanyRankingItemSchema"
            },
            "type": "array",
            "title": "Items",
            "description": "\u4e00\u81f4\u3057\u305f\u4f01\u696d\u4e00\u89a7"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "\u7dcf\u4ef6\u6570"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "description": "\u73fe\u5728\u30da\u30fc\u30b8\u756a\u53f7\uff081\u59cb\u307e\u308a\uff09"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size",
            "description": "1\u30da\u30fc\u30b8\u3042\u305f\u308a\u306e\u4ef6\u6570"
          },
          "search_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Summary",
            "description": "\u81ea\u7136\u8a00\u8a9e\u691c\u7d22\u6642\u306e\u89e3\u91c8\u8aac\u660e\uff08\u30d0\u30ca\u30fc\u8868\u793a\u7528\uff09"
          },
          "extracted_filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchConditionSchema"
              },
              {
                "type": "null"
              }
            ],
            "description": "\u81ea\u7136\u8a00\u8a9e\u691c\u7d22\u6642\u306b\u62bd\u51fa\u3055\u308c\u305f\u6761\u4ef6"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "page_size"
        ],
        "title": "CompanySearchResponse",
        "description": "\u4f01\u696d\u691c\u7d22\u30ec\u30b9\u30dd\u30f3\u30b9\uff08\u30e9\u30f3\u30ad\u30f3\u30b0\u4e00\u89a7 + \u81ea\u7136\u8a00\u8a9e\u691c\u7d22\u306e\u8981\u7d04\u4ed8\u304d\uff09.\n\n\u5b8c\u5168\u4e00\u81f4\u30fb\u901a\u5e38\u30d5\u30a3\u30eb\u30bf\u306e\u5834\u5408 search_summary / extracted_filters \u306f None\u3002"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportGenerateRequest": {
        "properties": {
          "format": {
            "const": "pdf",
            "title": "Format",
            "description": "\u51fa\u529b\u30d5\u30a9\u30fc\u30de\u30c3\u30c8\uff08\u73fe\u72b6 pdf \u306e\u307f\u5bfe\u5fdc\uff09",
            "default": "pdf"
          }
        },
        "type": "object",
        "title": "ReportGenerateRequest",
        "description": "\u6839\u62e0\u30ec\u30dd\u30fc\u30c8\u751f\u6210\u30ea\u30af\u30a8\u30b9\u30c8."
      },
      "ReportGenerateResponse": {
        "properties": {
          "file_id": {
            "type": "string",
            "format": "uuid",
            "title": "File Id",
            "description": "\u751f\u6210\u30d5\u30a1\u30a4\u30ebID"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed"
            ],
            "title": "Status",
            "description": "\u751f\u6210\u30b9\u30c6\u30fc\u30bf\u30b9"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url",
            "description": "\u30c0\u30a6\u30f3\u30ed\u30fc\u30c9URL\uff08completed \u6642\u306e\u307f\u8a2d\u5b9a\uff09"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "\u751f\u6210\u30ea\u30af\u30a8\u30b9\u30c8\u65e5\u6642"
          }
        },
        "type": "object",
        "required": [
          "file_id",
          "status",
          "created_at"
        ],
        "title": "ReportGenerateResponse",
        "description": "\u6839\u62e0\u30ec\u30dd\u30fc\u30c8\u751f\u6210\u958b\u59cb\u30ec\u30b9\u30dd\u30f3\u30b9."
      },
...
      "SearchConditionSchema": {
        "properties": {
          "unrealized_gain_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unrealized Gain Min",
            "description": "\u542b\u307f\u76ca\u4e0b\u9650\uff08\u5104\u5186\uff09"
          },
          "unrealized_gain_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unrealized Gain Max",
            "description": "\u542b\u307f\u76ca\u4e0a\u9650\uff08\u5104\u5186\uff09"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region",
            "description": "\u5730\u57df\uff08\u4f8b\uff1a\u95a2\u897f\uff09"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry",
            "description": "\u696d\u7a2e\uff08\u4f8b\uff1a\u5c0f\u58f2\uff09"
          },
          "pbr_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pbr Max",
            "description": "PBR \u4e0a\u9650\uff08\u4f8b\uff1a1.0\uff09"
          },
          "pbr_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pbr Min",
            "description": "PBR \u4e0b\u9650"
          },
          "structure_score_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structure Score Min",
            "description": "\u69cb\u9020\u30b9\u30b3\u30a2\u4e0b\u9650"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name",
            "description": "\u4f01\u696d\u540d\uff08\u5b8c\u5168\u4e00\u81f4\uff09"
          },
          "securities_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Securities Code",
            "description": "\u8a3c\u5238\u30b3\u30fc\u30c9\uff08\u5b8c\u5168\u4e00\u81f4\uff09"
          }
        },
        "type": "object",
        "title": "SearchConditionSchema",
        "description": "\u81ea\u7136\u8a00\u8a9e\u30af\u30a8\u30ea\u304b\u3089\u62bd\u51fa\u3055\u308c\u305f\u691c\u7d22\u6761\u4ef6\uff08\u30d5\u30a3\u30eb\u30bf\uff09.\n\n\u5168\u30d5\u30a3\u30fc\u30eb\u30c9 optional\u3002\u672a\u62bd\u51fa\u306e\u30d5\u30a3\u30fc\u30eb\u30c9\u306f None\u3002"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}